apps/orchestration/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.1
langchain-anthropic==0.1.11
langchain-community==0.2.1
langgraph==0.1.5
langsmith==0.1.57
tavily-python==0.3.3
anthropic==0.26.0
openai==1.30.1
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
